+++ apps/web/lib/qr/types.ts
@@ -1,0 +1,29 @@
+import qrcodegen from "./codegen";
+import type { CSSProperties } from "react";
+
+export type Modules = ReturnType<qrcodegen.QrCode["getModules"]>;
+export type Excavation = { x: number; y: number; w: number; h: number };
+
+export type ImageSettings = {
+  src: string;
+  height: number;
+  width: number;
+  excavate: boolean;
+  x?: number;
+  y?: number;
+};
+
+export type QRProps = {
+  value: string;
+  size?: number;
+  // Should be a real enum, but doesn't seem to be compatible with real code.
+  level?: string;
+  bgColor?: string;
+  fgColor?: string;
+  style?: CSSProperties;
+  includeMargin?: boolean;
+  imageSettings?: ImageSettings;
+};
+export type QRPropsCanvas = QRProps &
+  React.CanvasHTMLAttributes<HTMLCanvasElement>;
+export type QRPropsSVG = QRProps & React.SVGProps<SVGSVGElement>;
